Completed css flex exercise 05-flex-modal
This one gave me problems with how I was editing the file - with all of
my changes at the bottom so I could separate my work from the starting
template. Turns out, the cascading of CSS made it so that even with all
of the correct properties, I was not getting the desired result. Started
over editing the CSS inline w/ the template (adding properties to the
existing selectors rather than trying to create all new selectors) and
this ended up being much simpler than I had thought.